backend/requirements.txt
# EditScribe Backend Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM Clients (OpenRouter uses OpenAI client)
openai==1.54.0

# Document Processing
python-docx==1.1.0
PyPDF2==3.0.1
markdown==3.5.1

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
rich==13.7.0

# Testing
pytest==7.4.3
httpx==0.25.2

# Retry logic
tenacity==8.2.3
